Refactored and generalized protected route component
